Added Movelist scraper
Updated the program to pull the Move List instead. Accomplishes fundamentally the same task, but is (in my opinion) more readable and follows better coding practices.

If I were to improve the code, it would be to make Part3(Step2) of PullMoveList() thread safe and have the pull request for each move be handled by worker threads.